Implemented PlacementDecision calculation and handling logic, alongside a few fixes
This implementation may not be the optimal solution for the placement-decision maintenance problem, but it is an efficient enough set of steps towards that, given the existing code-base and controller workflow as grounding.

The implementation:
- Fixes a race-condition that was present in the use of gvkMapper, while also introducing gvr-mapping
- Adds a supporting struct: PlacementDecisionResolver that maintains the resolutions of all placement-decisions
- Adds handling for placement-decision objects, which includes handling out-of-sync objects and creating/updating objects when a decision is changed.
- Marks code-pieces to be deleted when Nir's transport code is ready

The flow is as follows:
an object changes -> placement-decision resolutions are updated -> modified placement-decisions are enqueued to be synced by workers -> new spec is calculated and patched for each updated resolution.

====

Rebased onto main then squashed commits made during code review with MikeSpreitzer.